+++ core/test/test_vpn_manager_fsm.cpp
@@ -249,6 +249,7 @@ static void vpn_handler(void *arg, VpnEvent what, void *data) {
     case VPN_EVENT_DNS_UPSTREAM_UNAVAILABLE:
     case VPN_EVENT_TUNNEL_CONNECTION_STATS:
     case VPN_EVENT_TUNNEL_CONNECTION_CLOSED:
+    case VPN_EVENT_CONNECTION_INFO:
         break;
     }
 }